move AppProcessesConfig schema to commcare_cloud/environment
